editor: remove intermediate org cursor wrap
Instead of interfacing with the org node content through an additional
proxy type ~OrgNodeTextWrap~ the interface now returns Q_GADGET directly
wrapping org node item.
